requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
attrs==21.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk.cloud-assembly-schema==1.127.0
    # via
    #   aws-cdk.core
    #   aws-cdk.cx-api
aws-cdk.core==1.127.0
    # via devops-demo (setup.py)
aws-cdk.cx-api==1.127.0
    # via aws-cdk.core
aws-cdk.region-info==1.127.0
    # via aws-cdk.core
cattrs==1.8.0
    # via jsii
constructs==3.3.161
    # via aws-cdk.core
jsii==1.38.0
    # via
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
python-dateutil==2.8.2
    # via jsii
six==1.16.0
    # via python-dateutil
typing-extensions==3.10.0.2
    # via jsii
